pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "dj-database-url>=3.1.0",
    "django>=5.2.10",
    "django-cors-headers>=4.9.0",
    "djangorestframework>=3.16.1",
    "openai>=2.14.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.1",
]
